# performance: keep trie prefetch during validation phase