@brilliant-lx 2023-11-02
reference: https://github.com/bnb-chain/bsc/pull/952